Add bc chacha20-ietf
Drop NaCl
