+++ Shadowsocks.Net/Shadowsocks.Net.csproj
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <PackageReference Include="BouncyCastle.NetCore" Version="1.8.6" />
-    <PackageReference Include="NaCl.Core" Version="2.0.0" />
     <PackageReference Include="Splat" Version="9.6.1" />
   </ItemGroup>
